requirements.txt
beautifulsoup4==4.6.0
bleach==2.1.1
certifi==2017.7.27.1
chardet==3.0.4
Django==1.10.8
django-contrib-comments==1.8.0
#django-storages==1.6.5
django-storages-redux
psycopg2
filebrowser-safe==0.4.7
future==0.16.0
grappelli-safe==0.4.7
html5lib==1.0b10
idna==2.6
Mezzanine==4.2.3
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tzlocal==1.4
urllib3==1.22
webencodings==0.5.1
zappa
django-compressor
boto
djangorestframework==3.7.3
